[Denis Çerri]Add Update Functionality
The project is split into small chunks for each contributor. My task was Update.\n When other contributors finish their task, i will add a commit with their name in the commit message.\n ToDo: \n - Check if the updated entity is trying to be something that is already in the database. Keys: Name, Surname