Make index of table rows start from 1
Since id is auto_increment, you will see jumps from one id to another. To fix that we create a temporary variable that grows every row that is added.
IF you sort in descending order, we first get the number of rows that are about to be put and count down from it.
Idea applied in index, search function in read.php and sort function in read.php